# ast: Remove noop walk visitor invocations.