Notable improvements are:

benchmark                                        old ns/op      new ns/op      delta
BenchmarkRESTAuthzForbidAuthn-8                  531944         481911         -9.41%
BenchmarkRESTAuthzForbidPath-8                   605428         559036         -7.66%
BenchmarkRESTAuthzForbidMethod-8                 615807         572775         -6.99%
BenchmarkRESTAuthzAllow10Paths-8                 608273         565498         -7.03%
BenchmarkRESTAuthzAllow100Paths-8                1203957        1126383        -6.44%
BenchmarkRESTAuthzAllow1000Paths-8               6377612        6456720        +1.24%
BenchmarkPartialEvalCompile/1-8                  4200311        3797585        -9.59%
BenchmarkPartialEvalCompile/10-8                 6882760        5234766        -23.94%
BenchmarkPartialEvalCompile/100-8                60144772       45053413       -25.09%
BenchmarkPartialEvalCompile/1000-8               3361430791     3154722341     -6.15%
BenchmarkWalk/100-8                              1032722        995608         -3.59%
BenchmarkWalk/1000-8                             1156007        1080659        -6.52%
BenchmarkWalk/2000-8                             1230263        1155320        -6.09%
BenchmarkWalk/3000-8                             1329760        1242109        -6.59%
BenchmarkInliningFullScan/1000-8                 9033921        7069875        -21.74%
BenchmarkInliningFullScan/10000-8                98827870       77699599       -21.38%
BenchmarkInliningFullScan/300000-8               3015002855     2393599775     -20.61%

Signed-off-by: Teemu Koponen <koponen@styra.com>

<!--

Thanks for submitting a PR to OPA!

Before pressing 'Create pull request' please read the checklist below.

* All code changes should be accompanied with tests. If you are not
modifying any tests, just provide a short explanation of why updates
to tests are not necessary. In addition to helping catch bugs, tests
are extremely helpful in providing _context_ that explains how your
changes can be used.

* All changes to public APIs **must** be accompanied with
docs. Examples of public APIs include built-in functions,
config fields, and of course, exported Go types/functions/constants/etc.

* Commit messages should explain _why_ you made the changes, not what
you changed. Use active voice. Keep the subject line under 50
characters or so.

* All commits must be signed off by the author. If you are not
familiar with signing off, see CONTRIBUTING.md below.

For more information on contributing to OPA see:

* [CONTRIBUTING.md](https://github.com/open-policy-agent/opa/blob/master/CONTRIBUTING.md)
  for high-level contribution guidelines.

* [DEVELOPMENT.md](https://github.com/open-policy-agent/opa/blob/master/docs/devel/DEVELOPMENT.md)
  for development workflow and environment setup.

-->